// frontend/src/utils/types.ts
export type UserLogin = {
  email: string;
  password: string;
};

export type Job = {
  customer: string;
  jobNumber: number;
  category: string;
  inspector: string;
  date: string;
  time: string;
  siteAddress: string;
  status: string;
  description: string;
  id: string;
};